+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.2.8] - 2022-XX-XX
 - Fix issue [#204](https://github.com/intersystems/language-server/issues/204): Class parser interprets modulo operator as a CSP extension
+- Fix issue [#210](https://github.com/intersystems/language-server/issues/210): Add Intellisense for Property data type parameters
 
 ## [1.2.7] - 2021-12-17
 - Fix issue [#201](https://github.com/intersystems/language-server/issues/201): Contents of `<EXAMPLE>` HTML tags shown on one line in Intellisense tooltips